+++ lib/Padding.h
@@ -4,12 +4,13 @@
 #include <cstddef>
 
 struct Padding {
-    size_t top;
-    size_t right;
-    size_t bottom;
-    size_t left;
+    unsigned top;
+    unsigned right;
+    unsigned bottom;
+    unsigned left;
 
-    Padding(size_t top, size_t right, size_t bottom, size_t left): top(top), right(right), bottom(bottom), left(left) {}
+    Padding(unsigned top, unsigned right, unsigned bottom, unsigned left) : top(top), right(right), bottom(bottom),
+                                                                            left(left) {}
 };
 
 #endif //SOFT354_MPI_PADDING_H
